changed books price from decimal to double

+++ src/main/kotlin/pt/unl/fct/iadi/bookstore/domain/Book.kt
@@ -13,6 +13,6 @@ data class Book(
     val isbn: String,
     var title: String,
     var author: String,
-    var price: BigDecimal,
+    var price: Double,
     var image: String
 )
